requirements.txt
selenium
beautifulsoup4
apache-airflow
azure-storage-blob
pendulum
webdriver-manager
dotenv